# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QueueLess API",
    "description": "Predict waiting time at government offices + remote virtual tokens",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/desk/tokens": {
      "post": {
        "tags": [
          "Desk"
        ],
        "summary": "Desk Create Token",
        "operationId": "desk_create_token_v1_desk_tokens_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeskBookIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeskSlipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/desk/tokens/{token_id}/check-in": {
      "post": {
        "tags": [
          "Desk"
        ],
        "summary": "Desk Manual Check In",
        "operationId": "desk_manual_check_in_v1_desk_tokens__token_id__check_in_post",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/tick": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "Trigger Tick",
        "description": "POST /internal/tick \u2014 Scheduler entry point.\nProtected by shared internal secret. Runs sweeps and notification delivery.",
        "operationId": "trigger_tick_internal_tick_post",
        "parameters": [
          {
            "name": "X-Internal-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Tick Internal Tick Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/dev-token": {
      "post": {
        "tags": [
          "Internal",
          "DevAuth"
        ],
        "summary": "Get Dev Token",
        "description": "POST /internal/dev-token \u2014 Generate a signed JWT for a pre-built dev persona.\nDISABLED in production. Protected by X-Internal-Secret.\n\nReturns tokens for all dev personas in one call \u2014 use for the DevAuth login selector.",
        "operationId": "get_dev_token_internal_dev_token_post",
        "parameters": [
          {
            "name": "X-Internal-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dev Token Internal Dev Token Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeskBookIn": {
        "properties": {
          "office_id": {
            "type": "string",
            "title": "Office Id"
          },
          "service_id": {
            "type": "string",
            "title": "Service Id"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "NORMAL"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "citizen_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citizen Name"
          },
          "created_via": {
            "type": "string",
            "title": "Created Via",
            "default": "ASSISTED"
          },
          "override_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Reason"
          }
        },
        "type": "object",
        "required": [
          "office_id",
          "service_id"
        ],
        "title": "DeskBookIn"
      },
      "DeskSlipOut": {
        "properties": {
          "token": {
            "$ref": "#/components/schemas/TokenOut"
          },
          "printable_code": {
            "type": "string",
            "title": "Printable Code"
          },
          "qr_data": {
            "type": "string",
            "title": "Qr Data"
          }
        },
        "type": "object",
        "required": [
          "token",
          "printable_code",
          "qr_data"
        ],
        "title": "DeskSlipOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "office_id": {
            "type": "string",
            "title": "Office Id"
          },
          "service_id": {
            "type": "string",
            "title": "Service Id"
          },
          "business_date": {
            "type": "string",
            "title": "Business Date"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "display_code": {
            "type": "string",
            "title": "Display Code"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "priority_status": {
            "type": "string",
            "title": "Priority Status"
          },
          "created_via": {
            "type": "string",
            "title": "Created Via"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "beneficiary_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Name"
          },
          "counter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Id"
          },
          "counter_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Label"
          },
          "arrived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrived At"
          },
          "called_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Called At"
          },
          "grace_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grace Deadline"
          },
          "serving_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "last_eta_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Eta Minutes"
          },
          "last_eta_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Eta Reason"
          },
          "eta_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Low"
          },
          "eta_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta High"
          },
          "waiting_ahead": {
            "type": "integer",
            "title": "Waiting Ahead",
            "default": 0
          },
          "now_serving": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now Serving"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "office_id",
          "service_id",
          "business_date",
          "seq",
          "display_code",
          "state",
          "category",
          "priority_status",
          "created_via"
        ],
        "title": "TokenOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}